backend/requirements.txt
anthropic>=0.39.0
openai>=1.54.0
requests>=2.31.0
python-dotenv>=1.0.0
